use pytest for unit tests
