+++ doc/source/ray-observability/ray-distributed-debugger.rst
@@ -55,7 +55,7 @@ Create a file `job.py` with the following snippet. Add `breakpoint()` in the Ray
 .. literalinclude:: ./doc_code/ray-distributed-debugger.py
     :language: python
 
-Run your Ray app 
+Run your Ray app
 ~~~~~~~~~~~~~~~~
 
 Start running your Ray app.
@@ -98,7 +98,7 @@ Run a Ray task raised exception
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Run the same `job.py` file with an additional argument to raise an exception.
-    
+
 .. code-block:: bash
 
     python job.py raise-exception
@@ -112,7 +112,7 @@ When the app throws an exception:
 - The debugger freezes the task.
 - The terminal clearly indicates when the debugger pauses a task and waits for the debugger to attach.
 - The paused task is listed in the Ray Debugger extension.
-- Click the play icon next to the name of the paused task to attach the debugger and start debugging. 
+- Click the play icon next to the name of the paused task to attach the debugger and start debugging.
 
 .. image:: ./images/post-moretem.gif
     :align: center
