+++ python/ray/data/_internal/util.py
@@ -257,7 +257,7 @@ def _warn_on_high_parallelism(requested_parallelism, num_read_tasks):
         and num_read_tasks > available_cpu_slots * 4
         and num_read_tasks >= 5000
     ):
-        logger.warn(
+        logger.warning(
             f"{WARN_PREFIX} The requested parallelism of {requested_parallelism} "
             "is more than 4x the number of available CPU slots in the cluster of "
             f"{available_cpu_slots}. This can "
